refactor(domain): 전략 패턴 패키지 이동
domain > domain.linegenerator
