+++ src/main/java/ladder/domain/linegenerator/LineGenerator.java
@@ -1,8 +1,7 @@
-package ladder.domain;
+package ladder.domain.linegenerator;
 
 import java.util.List;
 
-//TODO 패키지 이동(LineGenerator 한 패키지로 이동)
 public interface LineGenerator {
     List<Boolean> getLine(int peopleNumber);
 }
