+++ src/main/java/ladder/domain/linegenerator/RandomLineGenerator.java
@@ -1,4 +1,6 @@
-package ladder.domain;
+package ladder.domain.linegenerator;
+
+import ladder.domain.linegenerator.LineGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
